Main/Appliance: Allow users to specify a different storage controller and/or controller port for hard disks when importing a VM. bugref:5027
Follow-up build-fix due to a missing cast when assigning an unsigned
integer to a signed integer (flagged by [-Werror,-Wsign-conversion]).

svn:sync-xref-src-repo-rev: r149598

+++ src/VBox/Main/src-server/ApplianceImplImport.cpp
@@ -1,4 +1,4 @@
-/* $Id: ApplianceImplImport.cpp 149597 2022-01-28 16:09:52Z brent.paulson@oracle.com $ */
+/* $Id: ApplianceImplImport.cpp 149598 2022-01-28 16:31:30Z brent.paulson@oracle.com $ */
 /** @file
  * IAppliance and IVirtualSystem COM class implementations.
  */
@@ -5747,7 +5747,7 @@ l_skipped:
 
                             if (uOrigControllerPortValue != uNewControllerPortValue)
                             {
-                                AD.lPort = uNewControllerPortValue;
+                                AD.lPort = (int32_t)uNewControllerPortValue;
                             }
                             if (uOrigControllerIndex != uTargetControllerIndex)
                             {
